refactor:将社交平台组件与社交平台模块独立
